update 29-12-2023 - first working flow
in this update I modified both the cst and dxf scripts to work together to create a dxf, run the cst simulation and save its results.
the dxf script now has several modes that enable it to run together or apart from the cst script and it allows randome or prearrange seed for testings.
a VERY IMPORTANT NOTE! for now, I manually rewrite the dxf import entries in the history list to account for the full path of the dxf files