error fix and for update
I fixes a minor bug regarding plotting the dxfs and updated the script to run for loops with reports on progress.
i also save now the pictures of ants in a dedicated folder for us to quicly glimpse at the geometries it creates.
I also added the suppress_prints option for cleaner outputs.